Wrap everything together.

+++ lotus-parser/src/items/integer_literal.rs
@@ -15,6 +15,6 @@ impl IntegerLiteral {
             _ => self.value.parse().unwrap()
         };
 
-        Some(Wasm::typed(Type::Float, Wat::const_i32(i32_value)))
+        Some(Wasm::typed(Type::Integer, Wat::const_i32(i32_value)))
     }
 }
